More refactoring. Task roundtrip seems to be broken though.
