+++ html/src/templates/workers.html
@@ -58,13 +58,13 @@
                 <th>Stop</th>
             </tr> 
         </thead>
-        <tbody ng-repeat="worker in data.workers track by worker.pid">
+        <tbody ng-repeat="worker in data.workers track by worker.id">
             <tr>
                 <th scope="row">{{ ($index+1) }}</th>
-                <td><a ng-href="{{baseUrl}}/worker/{{worker.pid}}">{{ worker.pid }}</a></td>
+                <td><a ng-href="{{baseUrl}}/worker/{{worker.id}}">{{ worker.pid }}</a></td>
                 <td>{{ worker.rawTags }}</td>
                 <td>{{ worker.startedTs | date:'yyyy/MM/dd HH:mm:ss a' }}</td>
-                <td><a ng-href="{{baseUrl}}/worker/{{worker.pid}}/logs">{{ worker.logfile }}</a></td>
+                <td><a ng-href="{{baseUrl}}/worker/{{worker.id}}/logs">{{ worker.logfile }}</a></td>
                 <td>{{ worker.checkInterval }} seconds</td>
                 <td>
                     <button ng-click="data.stopWorker(worker)" type="button" class="btn btn-danger btn-sm">Stop</button>
