+++ tasks/tasks.go
@@ -1,210 +1,17 @@
 package tasks
 
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
 	log "github.com/Sirupsen/logrus"
-	"github.com/spf13/cast"
 	"github.com/spf13/viper"
-	"github.com/turtlemonvh/blanket/lib"
 	"gopkg.in/mgo.v2/bson"
-	"io"
-	"io/ioutil"
 	"net/http"
 	"os"
-	"path"
-	"regexp"
-	"strings"
-	"time"
+	"os/exec"
+	"text/template"
 )
-
-const (
-	DEFAULT_TIMEOUT = 3600 // default timeout is 1 hour
-)
-
-// Task types are just a load of configuration loaded with viper with a few extra methods
-type TaskType struct {
-	ConfigFile        string // path to TOML config file on disk
-	LoadedTs          int64  // time loaded from disk
-	ConfigVersionHash string // md5 hash of the config file
-	Config            *viper.Viper
-}
-
-var validConfigfileName = regexp.MustCompile(`(\w*).toml`)
-
-// Read types from config directory
-func ReadTypes() ([]TaskType, error) {
-	typesDir := viper.GetString("tasks.typesPath")
-
-	// Grab entries out of directory
-	var taskTypes []TaskType
-	dirEntries, err := ioutil.ReadDir(typesDir)
-	if err != nil {
-		// FIXME: Handle this more elegantly
-		return nil, err
-	}
-
-	for _, dirEntry := range dirEntries {
-		filepath := path.Join(typesDir, dirEntry.Name())
-		tt, err := ReadTaskTypeFromFilepath(filepath)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"error":    err.Error(),
-				"filepath": filepath,
-			}).Error("Problem loading toml file")
-			continue
-		}
-		taskTypes = append(taskTypes, tt)
-	}
-
-	return taskTypes, nil
-}
-
-func ReadTaskTypeFromFilepath(filepath string) (TaskType, error) {
-	// Check that the file exists and is a TOML file
-	fi, err := os.Stat(filepath)
-	if err != nil {
-		return TaskType{}, err
-	}
-	if fi.IsDir() {
-		return TaskType{}, fmt.Errorf("Path points to a directory")
-	}
-	if !validConfigfileName.MatchString(filepath) {
-		return TaskType{}, fmt.Errorf("Not a valid TOML file: %s", filepath)
-	}
-
-	configFile, err := os.Open(filepath)
-	if err != nil {
-		// FIXME: Handle this more elegantly
-		return TaskType{}, err
-	}
-	defer configFile.Close()
-
-	tt, err := readTaskType(configFile)
-	if err != nil {
-		// FIXME: Handle this more elegantly
-		return TaskType{}, err
-	}
-	tt.ConfigFile = filepath
-	tt.Config.Set("name", strings.Split(path.Base(filepath), ".toml")[0])
-
-	// Ignore errors in finding checksum for now
-	cs, err := lib.Checksum(filepath)
-	if err == nil {
-		tt.ConfigVersionHash = cs
-	}
-	return tt, nil
-}
-
-func readTaskType(configFile io.Reader) (TaskType, error) {
-	tt := TaskType{}
-	tt.Config = viper.New()
-	tt.Config.SetConfigType("toml")
-	tt.Config.SetDefault("timeout", DEFAULT_TIMEOUT)
-
-	err := tt.Config.ReadConfig(configFile)
-	if err != nil {
-		tt.Config.Set("validationError", err.Error())
-	}
-
-	tt.LoadedTs = time.Now().Unix()
-
-	// Check that required fields are set
-	if tt.Config.GetString("command") == "" {
-		return tt, fmt.Errorf("TaskType config file is missing required field 'command'.")
-	}
-
-	return tt, nil
-}
-
-func (t *TaskType) String() string {
-	return fmt.Sprintf("%s [loaded=%d]", t.Config.GetString("name"), t.LoadedTs)
-}
-
-func (t *TaskType) GetName() string {
-	return t.Config.GetString("name")
-}
-
-func (t *TaskType) ToJSON() (string, error) {
-	ttSettings := t.Config.AllSettings()
-	ttSettings["loadedTs"] = t.LoadedTs
-	ttSettings["configFile"] = t.ConfigFile
-	ttSettings["versionHash"] = t.ConfigVersionHash
-	bts, err := json.Marshal(ttSettings)
-	return string(bts), err
-}
-
-// Return a map of default values, {name => value}
-func (t *TaskType) DefaultEnv() map[string]string {
-	defaultEnv := cast.ToSlice(t.Config.Get("environment.default"))
-	env := make(map[string]string)
-	for _, envVar := range defaultEnv {
-		ev := cast.ToStringMap(envVar)
-		evName := cast.ToString(ev["name"])
-		evValue := cast.ToString(ev["value"])
-		env[evName] = evValue
-	}
-	return env
-}
-
-func (t *TaskType) HasRequiredEnv() bool {
-	defaultEnv := cast.ToSlice(t.Config.Get("environment.required"))
-	return len(defaultEnv) != 0
-}
-
-// Return a map of default values, {name => type(string)}
-func (t *TaskType) RequiredEnv() map[string]string {
-	defaultEnv := cast.ToSlice(t.Config.Get("environment.required"))
-	env := make(map[string]string)
-	for _, envVar := range defaultEnv {
-		ev := cast.ToStringMap(envVar)
-		evName := cast.ToString(ev["name"])
-		evType := cast.ToString(ev["type"])
-		env[evName] = evType
-	}
-	return env
-}
-
-// Tasks inherit all the environment params of a tasktype + more
-func (t *TaskType) NewTask(childEnv map[string]string) (Task, error) {
-	taskType := t.Config.GetString("name")
-	taskId := bson.NewObjectId()
-
-	// Merge environment variables
-	// FIXME: Take any files and copy them into directory
-	// FIXME: This should happen at execution time
-	mixedEnv := t.DefaultEnv()
-	for k, v := range childEnv {
-		mixedEnv[k] = v
-	}
-
-	log.WithFields(log.Fields{
-		"taskId":    taskId.Hex(),
-		"mixedEnv":  mixedEnv,
-		"childEnv":  childEnv,
-		"parentEnv": t.Config.GetStringMapString("environment"),
-	}).Info("Environment variable mixing results for task")
-
-	log.WithFields(log.Fields{
-		"taskId":   taskId.Hex(),
-		"taskType": taskType,
-		"tags":     t.Config.GetStringSlice("tags"),
-	}).Info("Tag mixing results for task")
-
-	return Task{
-		Id:            taskId,
-		CreatedTs:     time.Now().Unix(),
-		LastUpdatedTs: time.Now().Unix(),
-		TypeId:        t.Config.GetString("name"),
-		TypeDigest:    "",
-		ResultDir:     path.Join(viper.GetString("tasks.resultsPath"), taskId.Hex()),
-		State:         "WAITING",
-		WorkerId:      "",
-		Progress:      0,
-		ExecEnv:       mixedEnv,
-		Tags:          t.Config.GetStringSlice("tags"),
-	}, nil
-}
 
 var (
 	ValidTaskStates         = []string{"WAITING", "CLAIMED", "RUNNING", "ERROR", "SUCCESS", "STOPPED", "TIMEDOUT"}
@@ -230,7 +37,7 @@ type Task struct {
 }
 
 func (t *Task) String() string {
-	return fmt.Sprintf("%d [%d]", t.TypeId, t.CreatedTs)
+	return fmt.Sprintf("%s %s [%d]", t.TypeId, t.Id.Hex(), t.CreatedTs)
 }
 
 func (t *Task) ToJSON() (string, error) {
@@ -238,18 +45,47 @@ func (t *Task) ToJSON() (string, error) {
 	return string(bts), err
 }
 
-// FIXME: Move some of the complexity out of worker and into here
-func (t *Task) Execute() error {
-	// Fetch the task type for this task
-	filepath := path.Join(viper.GetString("tasks.typesPath"), fmt.Sprintf("%s.toml", t.TypeId))
-	tt, err := ReadTaskTypeFromFilepath(filepath)
+// Get the command object used to run this task
+// Task type is passed in so the same config is used for every step
+// Maybe task type TOML should be copied when task is added so that it is saved
+func (t *Task) GetCmd(tt *TaskType) (*exec.Cmd, error) {
+	var cmd *exec.Cmd
+	var err error
+
+	// Evaluate template
+	tmpl, err := template.New("tasks").Parse(tt.Config.GetString("command"))
 	if err != nil {
-		return err
+		log.WithFields(log.Fields{
+			"err": err.Error(),
+		}).Error("problem parsing task type's 'command' parameter as go template")
+		return cmd, err
+	}
+	var cmdString bytes.Buffer
+	err = tmpl.Execute(&cmdString, t.ExecEnv)
+	if err != nil {
+		log.WithFields(log.Fields{
+			"err":    err.Error(),
+			"taskId": t.Id,
+		}).Error("error evaluating template for command")
+		return cmd, err
 	}
 
-	fmt.Println("tt", tt)
+	// FIXME: Don't just use bash; use python, zsh, etc
+	cmd = exec.Command("bash", "-c", cmdString.String())
 
-	return nil
+	// Modify execution environment with env variables
+	// e.g. http://craigwickesser.com/2015/02/golang-cmd-with-custom-environment/
+	env := os.Environ()
+	for k, v := range t.ExecEnv {
+		env = append(env, fmt.Sprintf("%s=%s", k, v))
+	}
+	cmd.Env = env
+
+	return cmd, nil
+}
+
+func (t *Task) GetTaskType() (*TaskType, error) {
+	return FetchTaskType(t.TypeId)
 }
 
 // Refresh information about this task by pulling from the blanket server
